# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: paperz.pk — school exam paper maker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List teachers in the school
      responses:
        "200":
          description: Teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Invite/add a teacher to the school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /teachers/{id}:
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherUpdate"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Remove a teacher
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /classes:
    get:
      operationId: listClasses
      tags: [taxonomy]
      summary: List classes
      responses:
        "200":
          description: Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [taxonomy]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    patch:
      operationId: updateClass
      tags: [taxonomy]
      summary: Update a class
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteClass
      tags: [taxonomy]
      summary: Delete a class
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /subjects:
    get:
      operationId: listSubjects
      tags: [taxonomy]
      summary: List subjects
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [taxonomy]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [taxonomy]
      summary: Update a subject
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSubject
      tags: [taxonomy]
      summary: Delete a subject
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /chapters:
    get:
      operationId: listChapters
      tags: [taxonomy]
      summary: List chapters
      parameters:
        - name: subjectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [taxonomy]
      summary: Create a chapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Chapter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{id}:
    patch:
      operationId: updateChapter
      tags: [taxonomy]
      summary: Update a chapter
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterUpdate"
      responses:
        "200":
          description: Updated chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteChapter
      tags: [taxonomy]
      summary: Delete a chapter
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /topics:
    get:
      operationId: listTopics
      tags: [taxonomy]
      summary: List topics
      parameters:
        - name: chapterId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
    post:
      operationId: createTopic
      tags: [taxonomy]
      summary: Create a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicInput"
      responses:
        "201":
          description: Topic created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{id}:
    patch:
      operationId: updateTopic
      tags: [taxonomy]
      summary: Update a topic
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicUpdate"
      responses:
        "200":
          description: Updated topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTopic
      tags: [taxonomy]
      summary: Delete a topic
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Role:
      type: string
      enum: [super_admin, school_admin, teacher]
...
    Teacher:
      type: object
      required: [id, name, email, role, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        status:
          type: string
          enum: [active, invited]
        createdAt:
          type: ["string", "null"]
    TeacherInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 3
    TeacherUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
    Class:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer
    ClassInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        orderIndex:
          type: integer
    ClassUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        orderIndex:
          type: integer
    Subject:
      type: object
      required: [id, classId, name]
      properties:
        id:
          type: integer
        classId:
          type: integer
        name:
          type: string
    SubjectInput:
      type: object
      required: [classId, name]
      properties:
        classId:
          type: integer
        name:
          type: string
          minLength: 1
    SubjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
    Chapter:
      type: object
      required: [id, subjectId, name]
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer
    ChapterInput:
      type: object
      required: [subjectId, name]
      properties:
        subjectId:
          type: integer
        name:
          type: string
          minLength: 1
        orderIndex:
          type: integer
    ChapterUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        orderIndex:
          type: integer
    Topic:
      type: object
      required: [id, chapterId, name]
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        name:
          type: string
    TopicInput:
      type: object
      required: [chapterId, name]
      properties:
        chapterId:
          type: integer
        name:
          type: string
          minLength: 1
    TopicUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
...