# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * paperz.pk — school exam paper maker API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  /** @minLength 1 */
  name?: string;
}